Add containerd content store for plugins
This uses a containerd client to get a content store and uses that
instead of using the local blob store.

The old blob store is kept for compatability. A future commit will
migrate blobs from the blob store to the content store.

Additionally, Windows doesn't currently use containerd, so the blob
store is still around for that... though this is hardly a requirement
since Windows does not support v2 plugins.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>