Use bind mount for binary targets
This bind mounts the build context into the build instead of copying it.
Some changes were requried to move the bundles dir outside the source
path since these mounts are read-only.
This change is only used if the `PREFIX` value is set during build.

Additionally makes the binary targets not depend on the dev image since
most of that is not required to build our binaries.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>