Update PluginGeter interface to take context
This touches a _lot_ of things because plugingetter is used (at least
eventually) for a lot of things.

In some places (especially where there are many callers) I've chosen to
use a placeholder `context.TODO()` rather than wire a context all up the
call stack in order to keep the commit (mostly) as small as possible
since it is already rather large.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>