Attach: Break apart trace span
This makes it so we don't have a single long-running span for attaches.
Now we create a new span every 60s and end the old span with some
metadata to say that there are more spans coming.

Why: Spans are not exported until they end. For endpoints like attach
which is often used for an indefinite amount of time the span is never
exported. This change allows us to continuously export tracing data.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>