Down with the sickness (AUTO_GOPATH)
Go has a nice tool to inspect the currently set (or default) `GOPATH`
with `go env GOPATH`.
We should do that rather than absolutely requiring people to manually
set `GOPATH` or use the `AUTO_GOPATH`.

This doesn't actually remove `AUTO_GOPATH`, people can still use it.
But when its not set and a `GOPATH` is not explicitly set in the env
this will assume the result of `go env GOPATH` is ok to use (or error
out like before if its empty for some reason).

Signed-off-by: Brian Goff <cpuguy83@gmail.com>